+++ src/plonetheme/derico/setuphandlers.py
@@ -105,6 +105,7 @@ class HiddenProfiles:
             "plonetheme.derico.upgrades:1010",
             "plonetheme.derico.upgrades:1011",
             "plonetheme.derico.upgrades:1012",
+            "plonetheme.derico.upgrades:1013",
         ]